requirements.txt
streamlit>=1.22.0
openai>=1.0.0
requests>=2.28.2
python-dotenv>=1.0.0
markdown>=3.4.3
python-wordpress-xmlrpc>=2.3
notion-client>=2.0.0
pillow>=9.5.0
retry>=0.9.2 